Binding `list`'s "load more" button to `tap` event
Tap event (handled by jQuery Mobile) is not perfect. It used to (when
binding to `click`), touching that area would fire the event even when
you moved/panned your finger away from the button. By using the `tap`
event, this seems to be resolved. BUT now that same thing is happening
when using a mouse -- `mousedown`ing on the element will fire the event
even when `mouseup`ing somewhere OFF the element.