Fixed pesky #view page error
What was actually happening is this: when you go to the view page and
get an error state, you display it. But in the process of emptying the
`data-role="content"` div, you just killed the container div! So when
you come back, you can't append to it. Now we have a two-page system --
one for "container" messages and another for full-page messages. Create
the `data-role="content"` like so:     <div data-role="content">
<div id="name-content"></div>       <div id="name-container">
Container static text         <div id="name-container-content"></div>
</div>